+++ HISTORY.rst
@@ -2,6 +2,27 @@
 History
 =======
 
+0.29.0 (unreleased)
+-------------------
+
+Announcements
+~~~~~~~~~~~~~
+* It was found that the `ExtremeValues` adjustment algorithm was not as accurate and stable as first thought. It is now hidden from `xclim.sdba` but can still be accessed via `xclim.sdba.adjustment`, with a warning. Work on improving the algorithm is ongoing, and a better implementation will be in a future version.
+
+New indicators
+~~~~~~~~~~~~~~
+* ``effective_growing_degree_days`` indice returns growing degree days using dynamic start and end dates for the growing season (based on Bootsma et al. (2005)). This has also been wrapped as an indicator.
+* ``qian_weighted_mean_average`` (based on Qian et al. (2010)) is offered as an alternate method for determining the start date using a weighted 5-day average (``method="qian"``). Can also be used directly as an indice.
+* ``cold_and_dry_days`` indicator returns the number of days where the mean daily temperature is below the 25th percentile and the mean daily precipitation is below the 25th percentile over period. Added as ``CD`` to ICCLIM module.
+* ``warm_and_dry_days`` indicator returns the number of days where the mean daily temperature is above the 75th percentile and the mean daily precipitation is below the 25th percentile over period. Added as ``WD`` to ICCLIM module.
+* ``warm_and_wet_days`` indicator returns the number of days where the mean daily temperature is above the 75th percentile and the mean daily precipitation is above the 75th percentile over period. Added as ``WW`` to ICCLIM module.
+* ``cold_and_wet_days`` indicator returns the number of days where the mean daily temperature is below the 25th percentile and the mean daily precipitation is above the 75th percentile over period. Added as ``CW`` to ICCLIM module.
+
+Bug fixes
+~~~~~~~~~~~~~~
+* Various bug fixes in bootstrapping:
+   - in ``percentile_bootstrap`` decorator, fix the popping of bootstrap argument to propagate in to the function call.
+   - in ``bootstrap_func``, fix some issues with the resampling frequency which was not working when anchored.
 
 0.28.1 (2021-07-29)
 -------------------